+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Change Log
 - Name server lookup now also considers 127.0.1.1 when trying to find a name server on localhost.
   This is convenient on some systems (Debian Linux) where 127.0.1.1 is often the address assigned
   to the local system via the hosts file.
+- fixed multiplex server shutdown sometimes hanging
 - The default instance mode when using @expose on the class and not using @behavior, is now also 'session'
   (was 'single').   Note that when you used @behavior with its default argument or only @expose on methods,
   the instance mode of the class already was 'session'.